+++ server/src/main/java/com/project/yozmcafe/service/auth/AuthService.java
@@ -1,9 +1,9 @@
 package com.project.yozmcafe.service.auth;
 
-import com.project.yozmcafe.controller.auth.MemberInfo;
 import com.project.yozmcafe.controller.auth.OAuthProvider;
 import com.project.yozmcafe.controller.dto.TokenResponse;
 import com.project.yozmcafe.domain.member.Member;
+import com.project.yozmcafe.domain.member.MemberInfo;
 import com.project.yozmcafe.domain.member.MemberRepository;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
